--- controllers/create-article.js
@@ -1,27 +1,0 @@
-const Article = require("../models/article");
-// создаём статью
-module.exports.createArticle = (req, res, next) => {
-  const { keyword, title, text, date, source, link, image } = req.body;
-  Article.create({
-    keyword,
-    title,
-    text,
-    date,
-    source,
-    link,
-    image,
-    owner: req.user._id,
-  })
-    .then((article) =>
-      res.status(201).send({
-        keyword: article.keyword,
-        title: article.title,
-        text: article.text,
-        date: article.date,
-        source: article.source,
-        link: article.link,
-        image: article.image,
-      })
-    )
-    .catch((err) => next(err));
-};
