+++ middlewares/logger.js
@@ -5,3 +5,8 @@ module.exports.requestLogger = expressWinston.logger({
   transports: [new winston.transports.File({ filename: "request.log" })],
   format: winston.format.json(),
 });
+// логгер ошибок
+module.exports.errorLogger = expressWinston.errorLogger({
+  transports: [new winston.transports.File({ filename: "error.log" })],
+  format: winston.format.json(),
+});
