--- routes/create-article.js
@@ -1,8 +1,0 @@
-const router = require("express").Router();
-
-const { articleCheckSchema } = require("../validation/article-check-schema");
-const { createArticle } = require("../controllers/create-article");
-
-router.post("/", articleCheckSchema, createArticle);
-
-module.exports = router;
