--- validation/joi-schemas/id-schema.js
@@ -1,3 +1,0 @@
-const { Joi } = require("celebrate");
-// схема id
-module.exports.idSchema = Joi.string().alphanum().length(24);
