--- validation/joi-schemas/link-schema.js
@@ -1,3 +1,0 @@
-const { Joi } = require("celebrate");
-// схема ссылки
-module.exports.linkSchema = Joi.string().required().uri();
