--- validation/joi-schemas/text-schema.js
@@ -1,3 +1,0 @@
-const { Joi } = require("celebrate");
-// схема текста
-module.exports.textSchema = Joi.string().required();
